program/src/test/java/com/magnet/MenuTest.java
package com.magnet;
// Generado por CodiumAI

import org.junit.Test;
import static org.junit.Assert.*;
import java.util.List;

public class MenuTest {


    // El menú puede agregar un nuevo plato a la base de datos
    @Test
    public void test_agregarNuevoPlatoABaseDeDatos() {
        // Preparar
        Plato plato = new Plato("Plato", 10.99, "Receta");
        Menu menu = new Menu() {};
    
        // Actuar
        menu.agregarPlato(plato);
        List<String> listaMenu = menu.getMenuFromDatabase();
    
        // Afirmar
        assertTrue(listaMenu.contains("Plato"));
    }

    // El menú puede recuperar la lista de platos de la base de datos
    @Test
    public void test_recuperarListaDePlatosDeBaseDeDatos() {
        // Preparar
        Menu menu = new Menu() {};
    
        // Actuar
        List<String> listaMenu = menu.getMenuFromDatabase();
    
        // Afirmar
        assertNotNull(listaMenu);
    }

    // El menú puede manejar una lista vacía de platos
    @Test
    public void test_manipularListaVaciaDePlatos() {
        // Preparar
        Menu menu = new Menu() {};
    
        // Actuar
        List<String> listaMenu = menu.getMenuFromDatabase();
    
        // Afirmar
        assertTrue(listaMenu.isEmpty());
    }

    // El menú puede manejar una sintaxis SQL inválida
    @Test
    public void test_manipularSintaxisSqlInvalida() {
        // Preparar
        Menu menu = new Menu() {};
    
        // Actuar
        List<String> listaMenu = menu.getMenuFromDatabase();
    
        // Afirmar
        assertNotNull(listaMenu);
    }

    // El menú puede manejar valores de entrada inválidos
    @Test
    public void test_manipularValoresDeEntradaInvalidos() {
        // Preparar
        Plato plato = new Plato(null, -10.99, null);
        Menu menu = new Menu() {};
    
        // Actuar
        menu.agregarPlato(plato);
        List<String> listaMenu = menu.getMenuFromDatabase();
    
        // Afirmar
        assertFalse(listaMenu.contains(null));
    }

    // El menú puede manejar credenciales de base de datos inválidas
    @Test
    public void test_manipularCredencialesDeBaseDeDatosInvalidas() {
        // Preparar
        Plato plato = new Plato("Plato", 10.99, "Receta");
        Menu menu = new Menu() {};
    
        // Actuar
        menu.agregarPlato(plato);
        List<String> listaMenu = menu.getMenuFromDatabase();
    
        // Afirmar
        assertNotNull(listaMenu);
    }
}